Revert "add a python pkg to ml and rename mlp-example to test"
