--- ml/test/add_path.py
@@ -1,6 +1,0 @@
-#!/usr/bin/env python3
-import os
-import sys
-sys.path.append(
-    os.path.join(os.path.dirname(os.path.abspath(__file__)),
-                 "../src"))
